Flask-CleanArchitecture-main/src/infrastructure/databases/mssql.py
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker
from config import Config
from infrastructure.databases.base import Base

DATABASE_URL = Config.DATABASE_URI
engine = create_engine(DATABASE_URL)
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

def init_mssql():
    Base.metadata.create_all(bind=engine)
