Flask-CleanArchitecture-main/src/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-Login>=0.6
Flask-WTF>=1.2
python-dotenv>=1.0
marshmallow>=3.21
marshmallow-sqlalchemy>=1.0
reportlab>=4.2
requests>=2.32
pyodbc>=5.1 ; platform_system=="Windows"
